orchestration/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-cloud-bigquery==3.21.0
requests==2.31.0
python-multipart==0.0.9
google-auth==2.29.0
